feat: 代码框表单增加AI工具 --story=130106980 # Reviewed, transaction id: 71904
